feat(fzf-picker): _fzf_worktree_remove を Rust 化 + lib を役割別モジュールに分解
Issue #402 の B 群 Rust 化 PR2/3。旧 _fzf_worktree_remove を fzf-picker の
2 つ目の bin として実装し、__fzf_worktree_list.fish を撤去。あわせてレビュー指摘
（lib が闇鍋・テストの参照性が悪い）を受け、共有 lib と E2E を再設計した。

- src/bin/fzf-worktree-remove.rs: リンク worktree を fzf で選び確認のうえ削除。
  削除対象の内側にいた場合だけ退避先（メイン worktree）パスを stdout 出力し shim
  が cd。git は現在の worktree を消せないため削除前に自プロセスも退避。
- lib を役割ごとに 1 ファイルへ分解（lib.rs は設計方針 doc + mod 宣言のみ）:
  - worktree.rs: Worktree 型 + list_worktrees（git 取得 IO）
  - parse.rs:    parse_worktrees（porcelain パース・純）+ ユニットテスト
  - format.rs:   picker_lines / removal_lines（fzf 行成形・純）+ ユニットテスト
  - launch.rs:   run_fzf / command_exists（fzf 起動・PATH 探索の IO）
  純ロジックはユニットテスト同居、IO 層は E2E で検証、という分担を明確化。
- E2E を bin ごとに分割（tests/e2e/{fzf_ghq_cd,fzf_worktree_remove}.rs）、mod.rs は
  検証内容の目次 + 共有ヘルパー。内部ロジックのテストは E2E に置かない方針。
- _fzf_worktree_remove.fish を shim 化。__fzf_worktree_list.fish を削除。

残り（別 PR）: cdabbr。

Refs #402, #388
